agents/requirements.txt
fastapi
uvicorn
mcp
langchain-google-genai
langchain-core
python-dotenv